# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "GoBE is a powerful backend, MCP, and API server with a comprehensive route system following GoBE patterns.",
        "title": "GoBE API",
        "termsOfService": "https://github.com/rafa-mori/gobe/terms",
        "contact": {
            "name": "GoBE API Support",
            "url": "https://github.com/rafa-mori/gobe",
            "email": "support@gobe.dev"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/discord/approvals/{id}/approve": {
            "post": {
                "description": "Approves a pending approval request",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "discord"
                ],
                "summary": "Approve request",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Request ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/discord/approvals/{id}/reject": {
            "post": {
                "description": "Rejects a pending approval request",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "discord"
                ],
                "summary": "Reject request",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Request ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
...
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
    "security": [
        {
            "OAuth2": []
        }
    ],
...
}